+++ crates/aptos/CHANGELOG.md
@@ -4,7 +4,8 @@ All notable changes to the Aptos CLI will be captured in this file. This project
 
 ## Unreleased
 - `aptos move fmt` formats move files inside the `tests` and `examples` directory of a package.
-- Adds `aptos update prover-dependencies`, which installs the dependency of Move prover, boogie, z3 and cvc5.
+- Add `aptos update prover-dependencies`, which installs the dependency of Move prover, boogie, z3 and cvc5.
+- Update the default version of `movefmt` to be installed from 1.0.4 to 1.0.5
 
 ## [4.2.3] - 2024/09/20
 - Fix the broken indexer in localnet in 4.2.2, which migrates table info from sycn to async ways.